Tagging as 5.0.1. First release for WoW 5.0 / Mists of Pandaria.
Add the ability to use custom macrotext for in-combat and no mount situations.

Split the UI Options into 3-sub panels:
  * Mounts (was the old main option panel)
  * Macro : Unavailable (enter a macro to run when no mount can be found)
  * Macro : Combat (enter a macro to run in combat instead of the default)

Use the Blizzard localized version of slash commands.

Build the default in-combat macro on the fly in case we learnt a new form
since logging in.

Add a few more localized phrases, and (unfortunately) add two custom texts
to describe the macros that are non-localized.

A number of code cleanups and optimisations.  Probably some new bugs too.